added base model configured simplecnn and vit
